feat: sepearte ship
